src/python/telegram_bot/requirements.txt
python-telegram-bot==13.15
pymongo==4.5.0
python-dotenv==1.0.0